\name{FootballStatCalc}
\alias{FootballStatCalc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Offensive Football Stat Calculator
}
\description{
This functionset is designed to help the user calculate football statistics using easy, fun and straight forward methods. The user must input their own statlines for this functionset to work.
}
\usage{
qbgetstats()
wrgetstats()
rbgetstats()
}


\author{
Brandon Dora
}


\examples{
\ > qbgetstats()
\ Enter quarterback name: Bob B
\ Enter number of attempts: 403
\ Enter number of completions: 342
\ Enter number of passing yards: 3954
\ Enter number of passing touchdowns: 26
\ Enter number of interceptions: 4
\ Enter number of games played: 16
\ [1] "quarterback: Bob B"
\ [1] "number of attempts: 403"
\ [1] "number of completions: 342"
\ [1] "passing completion percentage: 84.863523573201"
\ [1] "total passing yards: 3954"
\ [1] "yards per attempt: 9.81141439205955"
\ [1] "total passing touchdowns: 26"
\ [1] "total interceptions: 4"
\ [1] "quarterback rating: 131.053556658395"
\ [1] "games played: 16"
  }
}

